Fix php numeric literal syntax highlighting bugs.
In php 7.4, numeric literals are allowed to use `_` as a separator.
https://wiki.php.net/rfc/numeric_literal_separator#restrictions

Also, `2.0, 'some string with a dot.'` would be highlighted as a syntax error
due to the old value of phpFloatError using `[eE.].*\.`, which went
beyond the characters in a valid float.

This commit doesn't fix `1.2+3.` (still incorrectly highlighted as an error),
but fixes many other cases.