Add php_htmlInStrings=2 to fix highlighting of $start='<?php';
This is pretty annoying when you're working with code that reads php
files or generates the contents of php files or generates snippets to
`eval()`.

When php_htmlInStrings was set to any value,
it would cause the html syntax rules to be embedded inside of the string
contents.

However, php.vim extends html.vim by allowing the php start tag to be
included, meaning that this is parsed as `<?php';`, i.e. the start of a
new string literal.

Work around that by using a different set of rules

The default behavior may be changed in the future for constants other
than 2 or 3

Fixes #12